move errorHandler to components folder
